pyproject.toml
[tool.poetry]
name = "discord-event-bot"
version = "0.1.0"
description = ""
authors = ["Rapptz <rapptz@gmail.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.7"
"discord.py" = "^1.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
